MCLOUD-5953: Backport Redis patches (#46)
