+++ patches.json
@@ -244,6 +244,14 @@
     },
     "Fix load balancer issue": {
       ">=2.3.4 <2.3.6": "MCLOUD-5837__fix_filesystem_load_balancer_issue__2.3.4.patch"
+    },
+    "Cache Locking performance issue": {
+      "2.3.3": "MDVA-27538__fix_performance_issue_in_cache_locking_mechanism__2.3.3.patch",
+      ">=2.3.3-p1 <2.3.4": "MDVA-27538__fix_performance_issue_in_cache_locking_mechanism__2.3.3-p1.patch",
+      ">=2.3.4 <2.3.5": "MDVA-26795__fix_performance_issue_in_cache_locking_mechanism__2.3.4.patch"
+    },
+    "Large amount of Keys Block Cache": {
+      ">=2.3.1 <2.3.5": "MDVA-22950__large_amount_of_keys_block_cache__2.3.1.patch"
     }
   },
   "magento/module-paypal": {
