+++ src/Test/Functional/Acceptance/AcceptanceCest.php
@@ -10,49 +10,8 @@ namespace Magento\CloudPatches\Test\Functional\Acceptance;
 /**
  * @group php73
  */
-class AcceptanceCest
+class AcceptanceCest extends AbstractCest
 {
-    /**
-     * @var string
-     */
-    protected $edition = 'EE';
-
-    /**
-     * @param \CliTester $I
-     */
-    public function _before(\CliTester $I): void
-    {
-        $I->cleanupWorkDir();
-    }
-
-    /**
-     * @param \CliTester $I
-     * @param string $templateVersion
-     */
-    protected function prepareTemplate(\CliTester $I, string $templateVersion): void
-    {
-        $I->cloneTemplateToWorkDir($templateVersion);
-        $I->createAuthJson();
-        $I->createArtifactsDir();
-        $I->createArtifactCurrentTestedCode('patches', '1.0.99');
-        $I->addArtifactsRepoToComposer();
-        $I->addEceDockerGitRepoToComposer();
-        $I->addDependencyToComposer('magento/magento-cloud-patches', '1.0.99');
-        $I->addDependencyToComposer(
-            'magento/magento-cloud-docker',
-            $I->getDependencyVersion('magento/magento-cloud-docker')
-        );
-
-        if ($this->edition === 'CE') {
-            $version = $this->getVersionRangeForMagento($I);
-            $I->removeDependencyFromComposer('magento/magento-cloud-metapackage');
-            $I->addDependencyToComposer('magento/ece-tools', '^2002.1.0');
-            $I->addDependencyToComposer('magento/product-community-edition', $version);
-        }
-
-        $I->composerUpdate();
-    }
-
     /**
      * @param \CliTester $I
      * @return string
@@ -91,16 +50,8 @@ class AcceptanceCest
         return [
             ['templateVersion' => '2.3.3'],
             ['templateVersion' => '2.3.4'],
+            ['templateVersion' => '2.3.5'],
             ['templateVersion' => 'master'],
         ];
-    }
-
-    /**
-     * @param \CliTester $I
-     */
-    public function _after(\CliTester $I): void
-    {
-        $I->stopEnvironment();
-        $I->removeWorkDir();
     }
 }
