+++ tests/functional/_data/files/patches/patch.patch
@@ -1,0 +1,7 @@
+--- target_file.md	2018-01-26 16:04:11.000000000 -0600
++++ target_file.md	2018-01-26 16:04:45.000000000 -0600
+@@ -1 +1,3 @@
+-# Hello World
++# Hello Magento
++
++## Additional Info
