+++ tests/functional/_data/files/patches/target_file.md
@@ -1,0 +1,1 @@
+# Hello World
