+++ tests/functional/_data/files/patches/target_file_applied_patch.md
@@ -1,0 +1,3 @@
+# Hello Magento
+
+## Additional Info
